refactor(web-ui): step-kind Lume tokens + 500-line splits (#478)
Review fixups for the templates-v2 branch:

- Lume gate: the step-kind palette (agent/action/human + badge text) was
  hardcoded hex twice (ConductorCanvas, TemplatePreview). It now lives as
  --step-kind-* tokens in theme.css, consumed by both components through
  the shared stepKindColors.ts map — no hex in component code, no
  keep-in-sync comment.
- 500-line rule: SaveAsTemplateDialog (527) extracted its slot editor
  sections into SaveAsTemplateSlotEditors.tsx (388/197);
  conductor/page.tsx (604) extracted the Roles and emit-event sections
  into ConductorRolesSection.tsx and ConductorEmitSection.tsx (464).
  Behavior-preserving: role/emit errors and refetches still route through
  the page, and the double-fire guard is shared via a guard callback
  (prop-object mutation trips react-hooks/immutability).
